move all scores to score package
