+++ src/main/java/io/github/mzmine/datamodel/features/types/numbers/scores/LipidAnnotationMsMsScoreType.java
@@ -1,0 +1,48 @@
+/*
+ *  Copyright 2006-2020 The MZmine Development Team
+ *
+ *  This file is part of MZmine.
+ *
+ *  MZmine is free software; you can redistribute it and/or modify it under the terms of the GNU
+ *  General Public License as published by the Free Software Foundation; either version 2 of the
+ *  License, or (at your option) any later version.
+ *
+ *  MZmine is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even
+ *  the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
+ *  Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License along with MZmine; if not,
+ *  write to the Free Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301
+ *  USA
+ */
+
+package io.github.mzmine.datamodel.features.types.numbers.scores;
+
+import io.github.mzmine.datamodel.features.types.numbers.abstr.ScoreType;
+import java.text.DecimalFormat;
+import java.text.NumberFormat;
+import org.jetbrains.annotations.NotNull;
+
+public class LipidAnnotationMsMsScoreType extends ScoreType {
+
+  public LipidAnnotationMsMsScoreType() {
+    super();
+  }
+
+  @Override
+  public NumberFormat getFormatter() {
+    return new DecimalFormat("0.0");
+  }
+
+  @Override
+  public @NotNull String getHeaderString() {
+    return "MS/MS Score";
+  }
+
+  @NotNull
+  @Override
+  public final String getUniqueID() {
+    // Never change the ID for compatibility during saving/loading of type
+    return "lipid_msms_score";
+  }
+}
